matched with dowble to avd accuracy losses*

+++ src/main/java/com/fno/retail/model/Affiliate.java
@@ -4,14 +4,14 @@ import java.util.Date;
 
 public class Affiliate extends User {
 
-    private static final float DISCOUNT = 10;
+    private static final double DISCOUNT = 10;
 
     public Affiliate(String name, Date registrationDare, double bill) {
         super(name, registrationDare, bill);
     }
 
     @Override
-    public float getDiscountPercentage() {
+    public double getDiscountPercentage() {
         return DISCOUNT;
     }
 }
